src/delay_collector/pyproject.toml
[project]
name = "delay-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "tzdata>=2025.2",
]
